some short of backend work

+++ backend/models.py
@@ -4,7 +4,7 @@ from flask_sqlalchemy import SQLAlchemy
 
 db=SQLAlchemy()
 
-# First entity
+# Entity1 User
 class User_Info(db.Model):
     __tablename__="user_info"
     id=db.Column(db.Integer,primary_key=True)
@@ -14,7 +14,7 @@ class User_Info(db.Model):
     full_name=db.Column(db.String,nullable=False)
     address=db.Column(db.String,nullable=False)
     pin_code=db.Column(db.Integer,nullable=False)
-    tickets=db.relationship("Ticket",cascade="all,delete",backref="user_info",lazy=True) # User can access all of Iis tickets
+    tickets=db.relationship("Ticket",cascade="all,delete",backref="user_info",lazy=True) # User can access all of Its tickets
 
 
 # Entity2 Theater
@@ -28,14 +28,14 @@ class Theatre(db.Model):
     shows=db.relationship("Show",cascade="all,delete",backref="theatre",lazy=True) # Theater can access all of Its shows
 
 
-# Entity3 show table
+# Entity3 Show
 class Show(db.Model):
     __tablename__="show"
     id=db.Column(db.Integer,primary_key=True)
     name=db.Column(db.String,nullable=False)
     tags=db.Column(db.String,nullable=False)
     rating=db.Column(db.Integer,default=1)
-    tkt_price=db.Column(db.Integer,default=0)
+    tkt_price=db.Column(db.Float,default=0.0)
     date_time=db.Column(db.DateTime,nullable=False)
     theatre_id=db.Column(db.Integer, db.ForeignKey("theatre.id"),nullable=False)
     tickets=db.relationship("Ticket",cascade="all,delete",backref="show",lazy=True) # Shows can access Its tickets
